requirements.txt
asgiref==3.2.7
bson==0.5.8
dataclasses==0.6
Django==2.2.11
djangorestframework==3.11.0
djongo==1.3.1
dnspython==1.16.0
gunicorn==20.0.4
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
whitenoise==5.0.1
sqlparse==0.2.4
